account settings and login updated
